Modularisation and isStaffingAllowed functionality
